// nestjs-nextjs-ecommerce/api/src/modules/payments/dto/payment-response.dto.ts
import { ApiProperty } from '@nestjs/swagger';
export class PaymentResponseDto {
  @ApiProperty({
    description: 'Unique identifier for the payment',
    example: '1215645s454sdosd4s-454sd',
  })
  id: string;

  @ApiProperty({
    description: 'Associated order ID for the payment',
    example: 'order-123',
  })
  orderId: string;

  @ApiProperty({
    description: 'Amount paid for the order',
    example: 99.99,
  })
  amount: number;

  @ApiProperty({
    description: 'ID of the user who made the payment',
    example: 'user-456',
  })
  userId: string;

  @ApiProperty({
    description: 'Currency used for the payment',
    example: 'USD',
  })
  currency: string;

  @ApiProperty({
    description: 'Current status of the payment',
    example: 'COMPLETED',
    enum: ['PENDING', 'COMPLETED', 'FAILED', 'CANCELLED'],
  })
  status: string;

  @ApiProperty({
    description: 'Payment method used for the transaction',
    example: 'stripe',
    nullable: true,
  })
  paymentMethod: string | null;

  @ApiProperty({
    description: 'Transaction ID from the payment gateway',
    example: 'pi_121345678',
    nullable: true,
  })
  transactionId: string | null;

  @ApiProperty({
    description: 'Timestamp when the payment was created',
    example: '2023-10-01T12:34:56.789Z',
  })
  createdAt: Date;

  @ApiProperty({
    description: 'Timestamp when the payment was last updated',
    example: '2023-10-01T12:34:56.789Z',
  })
  updatedAt: Date;
}

export class CreatePaymentIntentResponse {
  // generated by stripe
  @ApiProperty({
    example: 'pi_1234567890abcdef',
    description: 'Stripe client secret for payment confirmation',
  })
  clientSecret: string;

  @ApiProperty({
    example: '216545-454-sds4s854d65',
    description: 'Payment ID in database',
  })
  paymentId: string;

  @ApiProperty({
    example: 'Payment retrieved successfully',
  })
  message?: string;
}

export class PaymentApiResponseDto {
  @ApiProperty({
    description: 'Indicates whether the payment operation was successful',
    example: true,
  })
  success: boolean;

  // Related to stripe
  @ApiProperty({
    type: CreatePaymentIntentResponse,
  })
  data: PaymentResponseDto; // define the shape of payment details returned by API

  @ApiProperty({
    example: 'Payment intent created successfully',
    required: false,
  })
  message?: string; // optional message field for additional information of the response
}